requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
werkzeug>=3.0.0
torch>=2.5.0
torchvision>=0.16.0
numpy>=1.24.0
scikit-learn>=1.3.0
pillow>=10.0.0
ipykernel>=6.25.0
requests>=2.31.0
beautifulsoup4>=4.12.0
gunicorn>=21.0.0
